+++ terracottaWithWiperdog/runWiperdog.sh
@@ -1,11 +1,13 @@
-echo "Terracotta server address:"
+echo "Terracotta server address [default localhost:9510]:"
         read TC_URL
         TC_URL=TC_URL | tr -d ' '
         if [ "$TC_URL" = "" ]; then
                 TC_URL=localhost:9510
         fi
-
-        mv -f wiperdog/etc/quartz.properties wiperdog/etc/quartz.properties_bak
+	
+	if [ ! -f wiperdog/etc/quartz.properties ];then
+          mv -f wiperdog/etc/quartz.properties wiperdog/etc/quartz.properties_bak
+	fi
         echo org.quartz.scheduler.instanceName:TestScheduler >> wiperdog/etc/quartz.properties
         echo org.quartz.scheduler.instanceId:groovy_instance >> wiperdog/etc/quartz.properties
         echo org.quartz.scheduler.skipUpdateCheck:true >> wiperdog/etc/quartz.properties
